Split layergroup controllers into small controllers
